Move compute_logprobs into util
