+++ src/forge/util/__init__.py
@@ -3,6 +3,7 @@
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
+from .compute_logprobs import compute_logprobs
 from .distributed import get_world_size_and_rank
 from .logging import get_logger, log_once, log_rank_zero
 from .metric_logging import get_metric_logger
@@ -13,4 +14,5 @@ __all__ = [
     "log_once",
     "log_rank_zero",
     "get_metric_logger",
+    "compute_logprobs",
 ]
