+++ tests/unit_tests/test_compute_logprobs.py
@@ -5,7 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 
 """
-Tests for reference_actor.py - compute_logprobs function
+Tests for compute_logprobs function
 """
 
 import unittest
@@ -13,27 +13,14 @@ import unittest
 import pytest
 import torch
 
-
-def _import_error():
-    try:
-        import forge.actors.reference_model  # noqa: F401
-
-        return False
-    except Exception:
-        return True
+from forge.util import compute_logprobs
 
 
 class TestComputeLogprobs(unittest.TestCase):
     """Test the compute_logprobs utility function."""
 
-    @pytest.mark.skipif(
-        _import_error(),
-        reason="Import error, likely due to missing dependencies on CI.",
-    )
     def test_compute_logprobs_basic(self):
         """Test basic logprobs computation."""
-        from forge.actors.reference_model import compute_logprobs
-
         batch_size = 1
         seq_len = 5
         vocab_size = 1000
@@ -51,14 +38,8 @@ class TestComputeLogprobs(unittest.TestCase):
         assert result.shape == (batch_size, response_len)
         assert torch.all(result <= 0)  # Log probabilities should be <= 0
 
-    @pytest.mark.skipif(
-        _import_error(),
-        reason="Import error, likely due to missing dependencies on CI.",
-    )
     def test_compute_logprobs_with_temperature(self):
         """Test logprobs computation with temperature scaling."""
-        from forge.actors.reference_model import compute_logprobs
-
         batch_size = 1
         seq_len = 5
         vocab_size = 1000
@@ -76,14 +57,8 @@ class TestComputeLogprobs(unittest.TestCase):
         default_result = compute_logprobs(logits, input_ids)
         assert not torch.allclose(result, default_result)
 
-    @pytest.mark.skipif(
-        _import_error(),
-        reason="Import error, likely due to missing dependencies on CI.",
-    )
     def test_compute_logprobs_single_token(self):
         """Test logprobs computation with single token response."""
-        from forge.actors.reference_model import compute_logprobs
-
         batch_size = 1
         seq_len = 5
         vocab_size = 1000
@@ -97,14 +72,8 @@ class TestComputeLogprobs(unittest.TestCase):
         assert result.shape == (batch_size, response_len)
         assert result.numel() == 1  # Single element
 
-    @pytest.mark.skipif(
-        _import_error(),
-        reason="Import error, likely due to missing dependencies on CI.",
-    )
     def test_compute_logprobs_empty_response(self):
         """Test logprobs computation with empty response."""
-        from forge.actors.reference_model import compute_logprobs
-
         batch_size = 1
         seq_len = 5
         vocab_size = 1000
@@ -117,14 +86,8 @@ class TestComputeLogprobs(unittest.TestCase):
 
         assert result.shape == (batch_size, response_len)
 
-    @pytest.mark.skipif(
-        _import_error(),
-        reason="Import error, likely due to missing dependencies on CI.",
-    )
     def test_compute_logprobs_empty_prompt(self):
         """Test logprobs computation with empty prompt."""
-        from forge.actors.reference_model import compute_logprobs
-
         batch_size = 1
         vocab_size = 1000
         prompt_len = 0
